+++ src/pdf/filt_predict_png.h
@@ -8,7 +8,7 @@ pdf_filt_predict_png_decode(uint8_t *buf,
                             uint8_t colors,
                             uint8_t bpc,
                             uint16_t columns,
-                            uint16_t rows
+                            size_t rows
                             );
 
 // Encode PNG predictors
@@ -18,7 +18,7 @@ pdf_filt_predict_png_encode(uint8_t *buf,
                             uint8_t colors,
                             uint8_t bpc,
                             uint16_t columns,
-                            uint16_t rows,
+                            size_t rows,
                             uint8_t predictor
                             );
