+++ src/pdf/filt_predict_tiff.h
@@ -8,7 +8,7 @@ pdf_filt_predict_tiff_decode(uint8_t *in,
                                uint8_t colors,
                                uint8_t bpc,
                                uint16_t columns,
-                               uint16_t rows
+                               size_t rows
                                );
 
 // Encode TIFF predictors
@@ -18,7 +18,7 @@ pdf_filt_predict_tiff_encode(uint8_t *in,
                              uint8_t colors,
                              uint8_t bpc,
                              uint16_t columns,
-                             uint16_t rows
+                             size_t rows
                              );
 
 #endif
